Update Ex 2 project files
