+++ Demos/02-custom-fields/package.json
@@ -11,20 +11,20 @@
     "test": "gulp test"
   },
   "dependencies": {
-    "@microsoft/sp-core-library": "1.7.0",
-    "@microsoft/sp-lodash-subset": "1.7.0",
-    "@microsoft/sp-office-ui-fabric-core": "1.7.0",
-    "@microsoft/sp-webpart-base": "1.7.0",
+    "@microsoft/sp-core-library": "1.7.1",
+    "@microsoft/sp-lodash-subset": "1.7.1",
+    "@microsoft/sp-office-ui-fabric-core": "1.7.1",
+    "@microsoft/sp-webpart-base": "1.7.1",
     "@types/es6-promise": "0.0.33",
     "@types/webpack-env": "1.13.1",
     "react": "^16.3.2",
     "react-dom": "^16.3.2"
   },
   "devDependencies": {
-    "@microsoft/sp-build-web": "1.7.0",
-    "@microsoft/sp-module-interfaces": "1.7.0",
-    "@microsoft/sp-tslint-rules": "1.7.0",
-    "@microsoft/sp-webpart-workbench": "1.7.0",
+    "@microsoft/sp-build-web": "1.7.1",
+    "@microsoft/sp-module-interfaces": "1.7.1",
+    "@microsoft/sp-tslint-rules": "1.7.1",
+    "@microsoft/sp-webpart-workbench": "1.7.1",
     "@types/chai": "3.4.34",
     "@types/mocha": "2.2.38",
     "@types/react": "^16.4.2",
