+++ Demos/02-custom-fields/src/controls/PropertyPaneContinentSelector/IPropertyPaneContinentSelectorInternalProps.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 import { IPropertyPaneCustomFieldProps } from '@microsoft/sp-webpart-base';
 import { IPropertyPaneContinentSelectorProps } from './IPropertyPaneContinentSelectorProps';
