+++ Demos/02-custom-fields/src/controls/PropertyPaneContinentSelector/IPropertyPaneContinentSelectorProps.ts
@@ -1,3 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+import { IDropdownOption } from 'office-ui-fabric-react/lib/components/Dropdown';
+
 export interface IPropertyPaneContinentSelectorProps {
   label: string;
   onPropertyChange: (propertyPath: string, newValue: any) => void;
