+++ Demos/02-custom-fields/src/controls/PropertyPaneContinentSelector/components/IContinentSelectorProps.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 import { IDropdownOption } from 'office-ui-fabric-react/lib/components/Dropdown';
 
 export interface IContinentSelectorProps {
