+++ Demos/02-custom-fields/src/controls/PropertyPaneContinentSelector/index.ts
@@ -1,3 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 export * from './IPropertyPaneContinentSelectorProps';
 export * from './IPropertyPaneContinentSelectorInternalProps';
-export * from './PropertyPaneContinentSelector';
+export * from './PropertyPaneContentSelector';
+
