+++ Demos/02-custom-fields/src/webparts/helloPropertyPane/HelloPropertyPaneWebPart.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 import { Version } from '@microsoft/sp-core-library';
 import {
   BaseClientSideWebPart,
@@ -7,18 +10,18 @@ import {
 } from '@microsoft/sp-webpart-base';
 import { escape } from '@microsoft/sp-lodash-subset';
 
-import styles from './HelloPropertyPaneWebPart.module.scss';
-import * as strings from 'HelloPropertyPaneWebPartStrings';
-
 import {
   PropertyPaneContinentSelector,
   IPropertyPaneContinentSelectorProps
 } from '../../controls/PropertyPaneContinentSelector';
 
+import styles from './HelloPropertyPaneWebPart.module.scss';
+import * as strings from 'HelloPropertyPaneWebPartStrings';
+
 export interface IHelloPropertyPaneWebPartProps {
   description: string;
   myContinent: string;
-  numContinentsVisited: number;
+  numContinentsVisited: number;  
 }
 
 export default class HelloPropertyPaneWebPart extends BaseClientSideWebPart<IHelloPropertyPaneWebPartProps> {
@@ -33,7 +36,7 @@ export default class HelloPropertyPaneWebPart extends BaseClientSideWebPart<IHel
               <p class="${ styles.subTitle }">Customize SharePoint experiences using Web Parts.</p>
               <p class="${ styles.description }">${escape(this.properties.description)}</p>
               <p class="${ styles.description }">Continent where I reside: ${escape(this.properties.myContinent)}</p>
-              <p class="${ styles.description }">Number of continents I've visited: ${this.properties.numContinentsVisited}</p>
+              <p class="${ styles.description }">Number of continents I've visited: ${this.properties.numContinentsVisited}</p>              
               <a href="https://aka.ms/spfx" class="${ styles.button }">
                 <span class="${ styles.label }">Learn more</span>
               </a>
@@ -46,6 +49,21 @@ export default class HelloPropertyPaneWebPart extends BaseClientSideWebPart<IHel
   protected get dataVersion(): Version {
     return Version.parse('1.0');
   }
+
+  private validateContinents(textboxValue: string): string {
+    const validContinentOptions: string[] = ['africa', 'antarctica', 'asia', 'australia', 'europe', 'north america', 'south america'];
+    const inputToValidate: string = textboxValue.toLowerCase();
+
+    return (validContinentOptions.indexOf(inputToValidate) === -1)
+      ? 'Invalid continent entry; valid options are "Africa", "Antarctica", "Asia", "Australia", "Europe", "North America", and "South America"'
+      : '';
+  }
+
+  private onContinentSelectionChange(propertyPath: string, newValue: any): void {
+    const oldValue: any = this.properties[propertyPath];
+    this.properties[propertyPath] = newValue;
+    this.render();
+  }  
 
   protected getPropertyPaneConfiguration(): IPropertyPaneConfiguration {
     return {
@@ -63,38 +81,23 @@ export default class HelloPropertyPaneWebPart extends BaseClientSideWebPart<IHel
                 }),
                 // PropertyPaneTextField('myContinent', {
                 //   label: 'Continent where I currently reside',
-                //   onGetErrorMessage: this.validateContinents.bind(this)
+                //   onGetErrorMessage: this.validateContinents.bind(this)                  
                 // }),
                 new PropertyPaneContinentSelector('myContinent', <IPropertyPaneContinentSelectorProps>{
                   label: 'Continent where I currently reside',
                   disabled: false,
                   selectedKey: this.properties.myContinent,
                   onPropertyChange: this.onContinentSelectionChange.bind(this),
-                }),
+                }),                
                 PropertyPaneSlider('numContinentsVisited', {
                   label: 'Number of continents I\'ve visited',
                   min: 1, max: 6, showValue: true,
-                })
+                })                  
               ]
             }
           ]
         }
       ]
     };
-  }
-
-  // private validateContinents(textboxValue: string): string {
-  //   const validContinentOptions: string[] = ['africa', 'antarctica', 'asia', 'north america', 'south america'];
-  //   const inputToValidate: string = textboxValue.toLowerCase();
-    
-  //   return (validContinentOptions.indexOf(inputToValidate) === -1)
-  //     ? 'Invalid continent entry; valid options are "Africa", "Antarctica", "Asia", "North America", and "South America"'
-  //     : '';
-  // }
-
-  private onContinentSelectionChange(propertyPath: string, newValue: any): void {
-    const oldValue: any = this.properties[propertyPath];
-    this.properties[propertyPath] = newValue;
-    this.render();
   }
 }
