--- Demos/02-custom-fields/teams/manifest.json
@@ -1,47 +1,0 @@
-{
-  "$schema": "https://developer.microsoft.com/en-us/json-schemas/teams/v1.2/MicrosoftTeams.schema.json",
-  "manifestVersion": "1.2",
-  "packageName": "HelloPropertyPane",
-  "id": "2c63754d-18aa-4856-912f-8c5888adb1af",
-  "version": "0.1",
-  "developer": {
-    "name": "SPFx + Teams Dev",
-    "websiteUrl": "https://products.office.com/en-us/sharepoint/collaboration",
-    "privacyUrl": "https://privacy.microsoft.com/en-us/privacystatement",
-    "termsOfUseUrl": "https://www.microsoft.com/en-us/servicesagreement"
-  },
-  "name": {
-    "short": "HelloPropertyPane"
-  },
-  "description": {
-    "short": "HelloPropertyPane description",
-    "full": "HelloPropertyPane description"
-  },
-  "icons": {
-    "outline": "tab20x20.png",
-    "color": "tab96x96.png"
-  },
-  "accentColor": "#004578",
-  "configurableTabs": [
-    {
-      "configurationUrl": "https://{teamSiteDomain}{teamSitePath}/_layouts/15/TeamsLogon.aspx?SPFX=true&dest={teamSitePath}/_layouts/15/teamshostedapp.aspx%3FopenPropertyPane=true%26teams%26componentId=2c63754d-18aa-4856-912f-8c5888adb1af",
-      "canUpdateConfiguration": false,
-      "scopes": [
-        "team"
-      ]
-    }
-  ],
-  "validDomains": [
-    "*.login.microsoftonline.com",
-    "*.sharepoint.com",
-    "*.sharepoint-df.com",
-    "spoppe-a.akamaihd.net",
-    "spoprod-a.akamaihd.net",
-    "resourceseng.blob.core.windows.net",
-    "msft.spoppe.com"
-  ],
-  "webApplicationInfo": {
-    "resource": "https://{teamSiteDomain}",
-    "id": "00000003-0000-0ff1-ce00-000000000000"
-  }
-}
